Distinguish bound socket type in the kernel
This patch modifies socket(2) passed type semantics. If SOCK_STREAM is passed,
TLS is used; if SOCK_DGRAM, DTLS is used. But SOCK_STREAM/SOCK_DGRAM do
not affect whether TCP/UDP socket is expected (or any other supported socket
type in the future). It enables DTLS over TCP as well. The type of bound socket
is handled transparently in the kernel space.

fixes: #15